fix2

+++ Assets/Scripts/LocationBar.cs
@@ -58,7 +58,7 @@ public class LocationBar : MonoBehaviour
     //Sets new position of StartBar in screen space
     public void SetStartBarPosition(Vector3 newPosition)
     {
-        Debug.Log("The start-loopbar has been set from beat " + (m_tokenPostion.GetTactPositionForLoopBarMarker(startBarPosition) + 1) + " to " + (m_tokenPostion.GetTactPositionForLoopBarMarker(newPosition) + 1) + ".");
+        Debug.Log("The start-loopbar has been set from beat " + (m_tokenPostion.GetTactPositionForLoopBarMarker(startBarPosition) + 1) + " to " + (int)(m_tokenPostion.GetTactPositionForLoopBarMarker(newPosition) + 1) + ".");
         this.startBarPosition = newPosition;
     }
